Optimize Realm API & Navigator
