+++ app/src/main/java/com/fstyle/structure_android/data/source/RepositoryModule.java
@@ -1,7 +1,6 @@
 package com.fstyle.structure_android.data.source;
 
 import com.fstyle.structure_android.data.source.local.realm.RealmApi;
-import com.fstyle.structure_android.data.source.local.realm.RealmImpl;
 import com.fstyle.structure_android.utils.dagger.AppScope;
 
 import dagger.Module;
@@ -17,6 +16,6 @@ public class RepositoryModule {
     @AppScope
     @Provides
     public RealmApi provideRealmApi() {
-        return new RealmImpl();
+        return new RealmApi();
     }
 }
