+++ app/src/main/java/com/fstyle/structure_android/data/source/local/realm/UserLocalDataSource.java
@@ -3,14 +3,11 @@ package com.fstyle.structure_android.data.source.local.realm;
 import android.support.annotation.NonNull;
 import com.fstyle.structure_android.data.model.User;
 import com.fstyle.structure_android.data.source.UserDataSource;
-import io.realm.Realm;
 import io.realm.RealmObject;
 import io.realm.RealmResults;
 import java.util.List;
 import javax.inject.Inject;
 import rx.Observable;
-import rx.Subscriber;
-import rx.functions.Action2;
 
 /**
  * Created by le.quang.dao on 14/03/2017.
@@ -28,7 +25,7 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
     @Override
     public void openTransaction() {
         if (mRealmApi == null) {
-            mRealmApi = new RealmImpl();
+            mRealmApi = new RealmApi();
         }
     }
 
@@ -44,70 +41,51 @@ public class UserLocalDataSource implements UserDataSource.LocalDataSource {
 
     @Override
     public Observable<Void> insertUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
-            @Override
-            public void call(Subscriber<? super Void> subscriber, Realm realm) {
-                realm.insert(user);
-                subscriber.onCompleted();
-            }
+        return mRealmApi.realmTransactionAsync((subscriber, realm) -> {
+            realm.insert(user);
+            subscriber.onCompleted();
         });
     }
 
     @Override
     public Observable<Void> updateUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
-            @Override
-            public void call(Subscriber<? super Void> subscriber, Realm realm) {
-                realm.insertOrUpdate(user);
-                subscriber.onCompleted();
-            }
-        });
+        return mRealmApi.realmTransactionAsync(((subscriber, realm) -> {
+            realm.insertOrUpdate(user);
+            subscriber.onCompleted();
+        }));
     }
 
     @Override
     public Observable<Void> deleteUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
-            @Override
-            public void call(Subscriber<? super Void> subscriber, Realm realm) {
-                RealmObject.deleteFromRealm(user);
-                subscriber.onCompleted();
-            }
+        return mRealmApi.realmTransactionAsync((subscriber, realm) -> {
+            RealmObject.deleteFromRealm(user);
+            subscriber.onCompleted();
         });
     }
 
     @Override
     public Observable<Void> insertOrUpdateUser(@NonNull final User user) {
-        return mRealmApi.realmTransaction(new Action2<Subscriber<? super Void>, Realm>() {
-            @Override
-            public void call(Subscriber<? super Void> subscriber, Realm realm) {
-                realm.insertOrUpdate(user);
-                subscriber.onCompleted();
-            }
+        return mRealmApi.realmTransactionAsync((subscriber, realm) -> {
+            realm.insertOrUpdate(user);
+            subscriber.onCompleted();
         });
     }
 
     @Override
     public Observable<List<User>> getAllUser() {
-        return mRealmApi.realmOnNewThread(new Action2<Subscriber<? super List<User>>, Realm>() {
-            @Override
-            public void call(Subscriber<? super List<User>> subscriber, Realm realm) {
-                RealmResults<User> users = realm.where(User.class).findAll();
-
-                subscriber.onNext(users);
-                subscriber.onCompleted();
-            }
+        return mRealmApi.realmGet((subscriber, realm) -> {
+            RealmResults<User> users = realm.where(User.class).findAll();
+            subscriber.onNext(users);
+            subscriber.onCompleted();
         });
     }
 
     @Override
     public Observable<User> getUserByUserLogin(final String userLogin) {
-        return mRealmApi.realmOnNewThread(new Action2<Subscriber<? super User>, Realm>() {
-            @Override
-            public void call(Subscriber<? super User> subscriber, Realm realm) {
-                User user = realm.where(User.class).equalTo("login", userLogin).findFirst();
-                subscriber.onNext(user);
-                subscriber.onCompleted();
-            }
+        return mRealmApi.realmGet((subscriber, realm) -> {
+            User user = realm.where(User.class).equalTo("login", userLogin).findFirst();
+            subscriber.onNext(user);
+            subscriber.onCompleted();
         });
     }
 }
