+++ app/src/main/java/com/fstyle/structure_android/screen/main/MainModule.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import com.fstyle.structure_android.data.source.UserRepository;
 import com.fstyle.structure_android.utils.dagger.ActivityScope;
 import com.fstyle.structure_android.utils.navigator.Navigator;
-import com.fstyle.structure_android.utils.navigator.NavigatorImpl;
 import com.fstyle.structure_android.utils.validator.Validator;
 import com.fstyle.structure_android.widget.dialog.DialogManager;
 import com.fstyle.structure_android.widget.dialog.DialogManagerImpl;
@@ -54,6 +53,6 @@ public class MainModule {
     @ActivityScope
     @Provides
     public Navigator provideNavigator() {
-        return new NavigatorImpl(mActivity);
+        return new Navigator(mActivity);
     }
 }
